removed as a platform component and made it a design-patter
