+++ src/platform/core/index.ts
@@ -111,13 +111,6 @@ export * from './search/search.module';
 import { CovalentStepsModule } from './steps/steps.module';
 export * from './steps/steps.module';
 
-/**
- * MANAGEMENT LIST
- */
-
-import { CovalentManagementListModule } from './management-list/management-list.module';
-export * from './management-list/management-list.module';
-
 @NgModule({
   imports: [
     HttpModule,
@@ -140,7 +133,6 @@ export * from './management-list/management-list.module';
     CovalentPagingModule.forRoot(),
     CovalentSearchModule.forRoot(),
     CovalentStepsModule.forRoot(),
-    CovalentManagementListModule.forRoot(),
   ],
   exports: [
     HttpModule,
@@ -163,7 +155,6 @@ export * from './management-list/management-list.module';
     CovalentPagingModule,
     CovalentSearchModule,
     CovalentStepsModule,
-    CovalentManagementListModule,
   ],
 })
 export class CovalentCoreModule {
